# centre-api/src/main/java/com/centre/poly/role/RoleController.java
package com.centre.poly.role;

import com.centre.poly.common.PageResponse;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("role")
@RequiredArgsConstructor
public class RoleController {

    public final RoleService roleService;

    @PreAuthorize("hasRole('ROLE_ADMIN')")
    @PostMapping
    public ResponseEntity<Integer> saveRole(@RequestBody @Valid RoleRequest request) {
       return ResponseEntity.ok().body(roleService.save(request));
    }

    @PreAuthorize("hasRole('ROLE_ADMIN')")
    @GetMapping
    public ResponseEntity<PageResponse<RoleResponse>> findAllRoles(
            @RequestParam(name = "page", defaultValue = "0",
                    required = false) int page,
            @RequestParam(name = "size", defaultValue = "10", required = false) int size) {
        return ResponseEntity.ok().body(roleService.findAll(page, size));
    }

}